+++ src/session-id-handler/SessionIdHandler.ts
@@ -1,0 +1,7 @@
+export default interface SessionIdHandler {
+  isSet: () => boolean
+
+  get: () => string
+
+  set: (sessionId: string) => void
+}