+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- use `window.sessionStorage` to save and retrieve user session id
+
 ### Deprecated
 
 ### Removed